service-workers/service-worker/navigation-preload/broken-chunked-encoding.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Navigation Preload with chunked encoding</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.sub.js"></script>
<script>
promise_test(async t => {
  const SCRIPT = 'resources/broken-chunked-encoding-worker.js';
  const SCOPE = 'resources/broken-chunked-encoding-scope.asis';

  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  const worker = registration.installing;
  await wait_for_state(t, worker, 'activated');
  const frame = await with_iframe(SCOPE);

  add_completion_callback(async() => {
    if (registration)
      await registration.unregister();
    if (frame)
      frame.remove();
  });

  assert_equals(frame.contentDocument.body.textContent,
                'PASS: preloadResponse resolved');
}, 'FetchEvent#preloadResponse resolves even if the body is sent with ' +
   'broken chunked encoding.');

promise_test(async t => {
  const SCRIPT = 'resources/broken-chunked-encoding-worker.js';
  const SCOPE = 'resources/chunked-encoding-scope.py?use_broken_body';

  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  const worker = registration.installing;
  await wait_for_state(t, worker, 'activated');
  const frame = await with_iframe(SCOPE);

  add_completion_callback(async() => {
    if (registration)
      await registration.unregister();
    if (frame)
      frame.remove();
  });

  assert_equals(frame.contentDocument.body.textContent,
                'PASS: preloadResponse resolved');
}, 'FetchEvent#preloadResponse resolves even if the body is sent with ' +
   'broken chunked encoding with some delays');

</script>
