LayoutTests/platform/glib/imported/w3c/web-platform-tests/service-workers/service-worker/navigation-preload/broken-chunked-encoding.https-expected.txt



FAIL FetchEvent#preloadResponse resolves even if the body is sent with broken chunked encoding. assert_equals: expected "PASS: preloadResponse resolved" but got "FAIL: preloadResponse rejected"
PASS FetchEvent#preloadResponse resolves even if the body is sent with broken chunked encoding with some delays
